artishine-backend/server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
firebase-admin==6.5.0
pydantic==2.8.2
python-multipart==0.0.9
